terraform/photography-app/lambda/generate_thumbnail/requirements.txt
Pillow==10.4.0
pillow-avif-plugin==1.4.6
